refactor retry and discard callbacks from ContentMessage to MessageWrapper
