+++ src/script/components/MessagesList/Message/MessageWrapper.tsx
@@ -25,6 +25,11 @@ import {container} from 'tsyringe';
 
 import {WebAppEvents} from '@wireapp/webapp-events';
 
+import {OutgoingQuote} from 'src/script/conversation/MessageRepository';
+import {ContentMessage} from 'src/script/entity/message/ContentMessage';
+import {Text} from 'src/script/entity/message/Text';
+import {MentionEntity} from 'src/script/message/MentionEntity';
+import {QuoteEntity} from 'src/script/message/QuoteEntity';
 import {t} from 'Util/LocalizerUtil';
 
 import {CallMessage} from './CallMessage';
@@ -88,6 +93,29 @@ export const MessageWrapper: React.FC<MessageParams & {hasMarker: boolean; isMes
       message.waitingButtonId(buttonId);
       messageRepository.sendButtonAction(conversation, message, buttonId);
     }
+  };
+
+  const onRetry = async (message: ContentMessage) => {
+    const firstAsset = message.getFirstAsset();
+
+    if (firstAsset instanceof Text) {
+      const messageId = message.id;
+      const messageText = firstAsset.text;
+      const mentions: MentionEntity[] = firstAsset.mentions() && firstAsset.mentions();
+      const incomingQuote = message.quote();
+
+      const isOutgoingQuote = (quoteEntity: QuoteEntity): quoteEntity is OutgoingQuote => {
+        return quoteEntity.hash !== undefined;
+      };
+      const quote: OutgoingQuote | undefined =
+        incomingQuote && isOutgoingQuote(incomingQuote) ? (incomingQuote as OutgoingQuote) : undefined;
+
+      await messageRepository.sendTextWithLinkPreview(conversation, messageText, mentions, quote, messageId);
+    }
+  };
+
+  const onDiscard = async () => {
+    await messageRepository.deleteMessageById(conversation, message.id);
   };
 
   const contextMenuEntries = ko.pureComputed(() => {
@@ -184,6 +212,8 @@ export const MessageWrapper: React.FC<MessageParams & {hasMarker: boolean; isMes
         onClickInvitePeople={onClickInvitePeople}
         onClickParticipants={onClickParticipants}
         onClickReceipts={onClickReceipts}
+        onDiscard={onDiscard}
+        onRetry={onRetry}
         isMessageFocused={isMessageFocused}
         isMsgElementsFocusable={isMsgElementsFocusable}
       />
